+++ projects/ngx-material/src/lib/ngx-side-menu/components/render-side-menu/render-side-menu.component.ts
@@ -37,4 +37,6 @@ export class RenderSideMenuComponent {
   toggleTab(tab: NgxSideMenuTabContext) {
     this.context.toggleTab(tab);
   }
+
+
 }