+++ projects/ngx-tools/src/lib/rendering/models/template-rendering.ts
@@ -37,7 +37,12 @@ export abstract class TemplateRendering<T extends {} = {}> {
     this.view = this.render(injector, this.context);
     this.view.detectChanges();
     this.view.markForCheck();
-    this.view.onDestroy(() => this.reset());
+    this.view.onDestroy(() => {
+      this.rendering?.nodes.filter(x => !!x.anchor)
+        .forEach(x => x.node.parentNode?.removeChild(x.node));
+      this.rendering = undefined;
+      this.view = undefined;
+    });
     return this.view;
   }
 
@@ -189,11 +194,6 @@ export abstract class TemplateRendering<T extends {} = {}> {
 
   detachChangeRef() {
     this.view?.detach();
-  }
-
-  private reset() {
-    this.view = undefined;
-    this.rendering = undefined;
   }
 }