+++ projects/ngx-material/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @juulsgaard/ngx-material
 
+## 17.1.4
+
+### Patch Changes
+
+- Remove destroy tracking for side menus
+- Updated dependencies
+  - @juulsgaard/ngx-tools@17.1.4
+
+## 17.1.3
+
+### Patch Changes
+
+- Handle source destruction for side-menus
+
 ## 17.1.2
 
 ### Patch Changes